Use full path to the extracted dll when launching the App
Previously the short name was passed, at which point corerun was (correctly) prepending the
working directory.  Previously this worked out of coincidence where TPA path was preferred
over app path.  That's sense been fixed.

We could probably remove the bundle dir argument as this would be inferred from the App
path, but I'm doing so at the moment.